UI edits, Slider Inputs, Flask Edits
Worked on some UI edits, and getting flask to take input from the
sliders. The flask functions reexport and edited Geojson file based on
the passed weights in python. I think I need to reload the geojson/use
ajax on a change, but inputs are not passing from the page to the app.
Yet. Still making progress.