Revise App to use Session
Start using session variables instead of globals, and change app
redirect to do a render template with new weights. Might require
resubmission, but the base url will work.
The geojson being served might be best stored as a session variable as
well (or its data). As it stands the geojson requires resubmission on
base refresh. This addresses most concerns however.