Use AJAX to update page without refresh
This change will make it so only the geojson refreshes on submission.
The downside to this is that the original url will keep the weights the
same on a per session basis. It might be prudent to add a reset button,
but for now, keeping the same weights might be considered a feature
rather than a bug. It will also reset once they close their browser
should they return.